requirements.txt
discord.py>=2.7
playwright>=1.40
certifi
